Moved AbstractModule#load() call from constructor to ItemAbstractModule#getModule(ItemStack)
Having the load() call in the constructor was causing problems initializing data in sub classes. Specifically when dealing with constructor hierachy setting data after load() was called. To fix this the call was moved to the getModule(ItemStack) method. Where it can be called after the module has been built. 

Honestly, moving it to the constructor was a bad idea to start with.